Remove NumberRange increment config deprecations
